// src/astro/models/HDChart.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { Chart } from './Chart';

export class HDChart {
    'design': Chart;
    'natal': Chart;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "design",
            "baseName": "design",
            "type": "Chart",
            "format": ""
        },
        {
            "name": "natal",
            "baseName": "natal",
            "type": "Chart",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return HDChart.attributeTypeMap;
    }

    public constructor() {
    }
}